+++ noether/catalogue/historic/scientific.py
@@ -5,7 +5,9 @@ Obscure scientific units.
 from math import pi
 from noether.core import Unit
 
-from ..scientific.si import hertz
+from noether.core.units import AffineUnit
+
+from ..scientific.si import hertz, kelvin
 from ..scientific.essential import nit
 
 
@@ -16,6 +18,16 @@ apostlib = blondel = Unit(nit / pi, "apostlib", "asb")
 skot = Unit(apostlib(1e-3), "skot", "sk")
 bril = Unit(apostlib(1e-7), "bril")
 lambert = Unit(apostlib(1e4), "lambert", "L")
+
 # TODO: foot-lambert
 
-# TODO: go over purple links :)
+# % Temperature units
+
+degN = degNewton = AffineUnit(
+    kelvin * 100/33, kelvin*273.15, "degNewton", "°N")
+degRe = Reaumur = AffineUnit(
+    kelvin*0.8, kelvin*273.15, "Réamur", "°Ré")
+degDe = Delisle = AffineUnit(
+    kelvin*-2/3, kelvin*373.15, "Delisle", "°De")
+degRo = Romer = AffineUnit(
+    kelvin * 40/21, kelvin * (273.15-7.5 * 40/21), "Rømer", "°Rø")